Fix the implementation of q_descend, q_ascend in queue.c
q_descend
The fix to original code utilize API from list.h in a
more concise way. The correctness of q_descend was
incorrect because the misunderstanding to the problem
description.
The correct understanding is as follows:
For every node, say node A, if there exists some nodes
whose value is larger than the node A, then delete this
node. Finally, return the size of the resulting queue.

q_ascend
The fix to original code utilize API from list.h in a
more concise way. The correctness of q_descend was
incorrect because the misunderstanding to the problem
description.
The correct understanding is as follows:
For every node, say node A, if there exists some nodes
whose value is less than the node A, then delete this
node. Finally, return the size of the resulting queue.

Change-Id: Id86ed675171488e3c773512043ae6c011eda8a89