Attempt to fix ERROR: Corruption detected in block
The error occurs when freeing queues that have
been merged by q_merge.

The structure is like this:
queue context type (points) head (points) ...
and in qtest.c it calls q_free on head, and
then free queue context type. I suspect the reason
behind this is because queue context type still
points to head after head is freed, so when I
call free on head in queue.c, harness.c detects
this issue, and gives the error.

For now, I can't come out a way to address this.

Change-Id: I68648bd32222b61b8ccd14597bd64b04e12bb10a